+++ mayan/apps/document_states/tests/mixins.py
@@ -168,9 +168,9 @@ class WorkflowStateViewTestMixin:
 
 
 class WorkflowTestMixin:
-    def _create_test_workflow(self, add_document_type=False):
+    def _create_test_workflow(self, add_document_type=False, auto_launch=True):
         self.test_workflow = Workflow.objects.create(
-            auto_launch=True, label=TEST_WORKFLOW_LABEL,
+            auto_launch=auto_launch, label=TEST_WORKFLOW_LABEL,
             internal_name=TEST_WORKFLOW_INTERNAL_NAME
         )
         self.test_workflow_runtime_proxy = WorkflowRuntimeProxy.objects.get(