+++ mayan/apps/document_states/views/workflow_template_state_views.py
@@ -57,7 +57,9 @@ class WorkflowTemplateStateActionCreateView(
         }
 
     def get_form_schema(self):
-        return self.get_class()().get_form_schema(request=self.request)
+        return self.get_class()().get_form_schema(
+            request=self.request, workflow_state=self.external_object
+        )
 
     def get_instance_extra_data(self):
         return {